requirements.txt
Flask==3.0.0
google-generativeai==0.5.4
python-dotenv==1.0.1